fix: properly patch cookies on bridge initialization
