+++ core/native-bridge.ts
@@ -2,6 +2,7 @@
  * Note: When making changes to this file, run `npm run build:nativebridge`
  * afterwards to build the nativebridge.js files to the android and iOS projects.
  */
+import { CapacitorCookies } from './src/core-plugins';
 import type {
   CallData,
   CapacitorInstance,
@@ -10,6 +11,7 @@ import type {
   PluginResult,
   WindowCapacitor,
 } from './src/definitions-internal';
+
 
 // For removing exports for iOS/Android, keep let for reassignment
 // eslint-disable-next-line
@@ -285,6 +287,58 @@ const initBridge = (w: any): void => {
 
       return String(msg);
     };
+
+    /**
+    * Safely web decode a string value (inspired by js-cookie)
+    * @param str The string value to decode
+    */
+    const decode = (str: string): string =>
+    str.replace(/(%[\dA-F]{2})+/gi, decodeURIComponent);
+
+    const platform = getPlatformId(win);
+
+    // patch document.cookie on Android/iOS
+    if (platform == 'android' || platform == 'ios') {
+      win.CapacitorCookiesDescriptor =
+        Object.getOwnPropertyDescriptor(Document.prototype, 'cookie') ||
+        Object.getOwnPropertyDescriptor(HTMLDocument.prototype, 'cookie');
+    
+      
+        Object.defineProperty(document, 'cookie', {
+          get: function () {
+            if (platform === 'ios') {
+              // Use prompt to synchronously get cookies.
+              // https://stackoverflow.com/questions/29249132/wkwebview-complex-communication-between-javascript-native-code/49474323#49474323
+  
+              const payload = {
+                type: 'CapacitorCookies'
+              }
+  
+              const res = prompt(JSON.stringify(payload));
+              return res;
+            }
+            else if(typeof win.CapacitorCookiesAndroidInterface !== 'undefined') {
+              return win.CapacitorCookiesAndroidInterface.getCookies();
+            }
+          },
+          set: function (val) {
+            const cookiePairs = val.split(';');
+            for (const cookiePair of cookiePairs) {
+              const cookieKey = cookiePair.split('=')[0];
+              const cookieValue = cookiePair.split('=')[1];
+
+              if (null == cookieValue) {
+                continue;
+              }
+
+              CapacitorCookies.setCookie({
+                key: cookieKey,
+                value: decode(cookieValue),
+              });
+            }
+          }
+        });
+    }
 
     // patch window.console on iOS and store original console fns
     const isIos = getPlatformId(win) === 'ios';
