+++ core/src/definitions-internal.ts
@@ -172,6 +172,8 @@ export interface CapacitorCustomPlatformInstance {
 
 export interface WindowCapacitor {
   Capacitor?: CapacitorInstance;
+  CapacitorCookiesAndroidInterface?: any;
+  CapacitorCookiesDescriptor?: PropertyDescriptor;
   /**
    * @deprecated Use `CapacitorCustomPlatform` instead
    */
